refactor(creation): Extract hooks and components from oversized cards (#236)
Extract reusable code from SkillsCard (2041→1405 lines) and
AugmentationsCard (1571→1030 lines) to improve maintainability.

SkillsCard extractions:
- useSkillDesignations: Free skill designation logic
- useGroupBreaking: Group break/restore with karma tracking
- useKarmaPurchase: Karma purchase mode detection
- utils.ts: Karma spent helpers

AugmentationsCard extractions:
- AugmentationItem: Non-cyberlimb display component
- CyberlimbAugmentationItem: Cyberlimb with mods display
- utils.ts: Shared formatting functions
