+++ components/creation/AugmentationsCard.tsx
@@ -26,11 +26,15 @@ import {
 } from "./shared";
 import {
   AugmentationModal,
+  AugmentationItem,
+  CyberlimbAugmentationItem,
   CyberwareEnhancementModal,
   CyberlimbAccessoryModal,
   CyberlimbWeaponModal,
+  formatCurrency,
+  formatEssence,
+  DISPLAY_CATEGORIES,
   type AugmentationSelection,
-  type AugmentationType,
   type CyberwareEnhancementSelection,
   type CyberlimbAccessorySelection,
   type CyberlimbWeaponSelection,
@@ -45,60 +49,8 @@ import {
   wouldReplaceExisting,
   isCyberlimb,
 } from "@/lib/types/cyberlimb";
-import {
-  Lock,
-  X,
-  AlertTriangle,
-  Cpu,
-  Heart,
-  Plus,
-  Zap,
-  ChevronDown,
-  ChevronRight,
-  Shield,
-} from "lucide-react";
+import { Lock, AlertTriangle, Cpu, Heart, Plus } from "lucide-react";
 import { InfoTooltip } from "@/components/ui";
-
-// =============================================================================
-// CONSTANTS
-// =============================================================================
-
-const GRADE_DISPLAY: Record<string, string> = {
-  used: "Used",
-  standard: "Std",
-  alpha: "Alpha",
-  beta: "Beta",
-  delta: "Delta",
-};
-
-/** Category groupings for display in the card */
-const DISPLAY_CATEGORIES = [
-  { id: "cyberlimb", label: "Cyberlimbs", type: "cyberware" },
-  { id: "headware", label: "Headware", type: "cyberware" },
-  { id: "eyeware", label: "Eyeware", type: "cyberware" },
-  { id: "earware", label: "Earware", type: "cyberware" },
-  { id: "bodyware", label: "Bodyware", type: "cyberware" },
-  { id: "cybernetic-weapon", label: "Cybernetic Weapons", type: "cyberware" },
-  { id: "basic", label: "Basic Bioware", type: "bioware" },
-  { id: "cultured", label: "Cultured Bioware", type: "bioware" },
-  { id: "cosmetic", label: "Cosmetic Bioware", type: "bioware" },
-] as const;
-
-// =============================================================================
-// HELPERS
-// =============================================================================
-
-function formatCurrency(value: number): string {
-  return new Intl.NumberFormat("en-US", {
-    style: "decimal",
-    minimumFractionDigits: 0,
-    maximumFractionDigits: 0,
-  }).format(value);
-}
-
-function formatEssence(value: number): string {
-  return value.toFixed(2);
-}
 
 // =============================================================================
 // TYPES
@@ -107,499 +59,6 @@ function formatEssence(value: number): string {
 interface AugmentationsCardProps {
   state: CreationState;
   updateState: (updates: Partial<CreationState>) => void;
-}
-
-// =============================================================================
-// AUGMENTATION ITEM COMPONENT
-// =============================================================================
-
-function AugmentationItem({
-  item,
-  type,
-  onRemove,
-  onAddEnhancement,
-  onRemoveEnhancement,
-}: {
-  item: CyberwareItem | BiowareItem;
-  type: "cyberware" | "bioware";
-  onRemove: () => void;
-  onAddEnhancement?: () => void;
-  onRemoveEnhancement?: (enhancementIndex: number) => void;
-}) {
-  const [isExpanded, setIsExpanded] = useState(false);
-  const isCyberware = type === "cyberware";
-  const cyberItem = item as CyberwareItem;
-  const hasCapacity = isCyberware && cyberItem.capacity && cyberItem.capacity > 0;
-  const remainingCapacity = hasCapacity
-    ? (cyberItem.capacity || 0) - (cyberItem.capacityUsed || 0)
-    : 0;
-  const enhancementCount = cyberItem.enhancements?.length || 0;
-
-  // Check if expandable (has bonuses, capacity, or enhancements)
-  const hasExpandableContent =
-    hasCapacity ||
-    enhancementCount > 0 ||
-    (item.attributeBonuses && Object.keys(item.attributeBonuses).length > 0) ||
-    item.armorBonus ||
-    (item as CyberwareItem).initiativeDiceBonus;
-
-  // Format item name with rating
-  const displayName = useMemo(() => {
-    // Check if name already contains rating info like "(Rating X)" and convert to RX
-    const ratingMatch = item.name.match(/\(Rating (\d+)\)/);
-    if (ratingMatch) {
-      return item.name.replace(/\s*\(Rating (\d+)\)/, ` R${ratingMatch[1]}`);
-    }
-    return item.name;
-  }, [item.name]);
-
-  // Format category display
-  const categoryDisplay = useMemo(() => {
-    if (!item.category) return "";
-    return item.category.charAt(0).toUpperCase() + item.category.slice(1).replace(/-/g, " ");
-  }, [item.category]);
-
-  return (
-    <div>
-      {/* Compact Single-Line Header */}
-      <div className="flex items-center gap-1.5 py-2">
-        {/* Expand/Collapse Button */}
-        {hasExpandableContent && (
-          <button
-            onClick={() => setIsExpanded(!isExpanded)}
-            aria-label={
-              isExpanded ? `Collapse ${displayName} details` : `Expand ${displayName} details`
-            }
-            aria-expanded={isExpanded}
-            className="shrink-0 text-zinc-400 hover:text-zinc-600 dark:hover:text-zinc-300"
-          >
-            {isExpanded ? (
-              <ChevronDown className="h-4 w-4" aria-hidden="true" />
-            ) : (
-              <ChevronRight className="h-4 w-4" aria-hidden="true" />
-            )}
-          </button>
-        )}
-
-        {/* Name */}
-        <span
-          className="text-sm font-medium text-zinc-900 dark:text-zinc-100 truncate flex-1"
-          title={displayName}
-        >
-          {displayName}
-        </span>
-
-        {/* Essence */}
-        <span
-          className={`text-sm font-medium shrink-0 ${
-            isCyberware ? "text-cyan-600 dark:text-cyan-400" : "text-pink-600 dark:text-pink-400"
-          }`}
-        >
-          {formatEssence(item.essenceCost)} ESS
-        </span>
-
-        {/* Cost */}
-        <span className="text-sm font-medium text-zinc-900 dark:text-zinc-100 shrink-0">
-          ¥{formatCurrency(item.cost)}
-        </span>
-
-        {/* Mod count badge (only if has mods) */}
-        {enhancementCount > 0 && (
-          <span className="text-[10px] text-zinc-500 dark:text-zinc-400 shrink-0">
-            [{enhancementCount} mod{enhancementCount !== 1 ? "s" : ""}]
-          </span>
-        )}
-
-        {/* Separator */}
-        <div className="mx-1 h-5 w-px bg-zinc-300 dark:bg-zinc-600" />
-
-        {/* Remove Button */}
-        <button
-          onClick={onRemove}
-          aria-label={`Remove ${displayName}`}
-          className="shrink-0 rounded p-1 text-zinc-400 hover:bg-red-100 hover:text-red-600 dark:hover:bg-red-900/30 dark:hover:text-red-400"
-        >
-          <X className="h-4 w-4" aria-hidden="true" />
-        </button>
-      </div>
-
-      {/* Expanded Details */}
-      {isExpanded && (
-        <div className="border-t border-zinc-100 dark:border-zinc-800 py-3 ml-6 space-y-3">
-          {/* Grade & Category */}
-          <div className="text-xs text-zinc-500 dark:text-zinc-400">
-            {GRADE_DISPLAY[item.grade]} • {categoryDisplay}
-          </div>
-
-          {/* Bonuses (Armor, Attributes, Initiative) */}
-          {((item.attributeBonuses && Object.keys(item.attributeBonuses).length > 0) ||
-            item.armorBonus ||
-            (item as CyberwareItem).initiativeDiceBonus) && (
-            <div className="flex flex-wrap gap-1">
-              {item.armorBonus && (
-                <span className="flex items-center gap-1 rounded bg-amber-100 px-1.5 py-0.5 text-[10px] font-medium text-amber-700 dark:bg-amber-900/50 dark:text-amber-300">
-                  <Shield className="h-3 w-3" />+{item.armorBonus} Armor
-                </span>
-              )}
-              {(item as CyberwareItem).initiativeDiceBonus && (
-                <span className="flex items-center gap-1 rounded bg-purple-100 px-1.5 py-0.5 text-[10px] font-medium text-purple-700 dark:bg-purple-900/50 dark:text-purple-300">
-                  <Zap className="h-3 w-3" />+{(item as CyberwareItem).initiativeDiceBonus}D6 Init
-                </span>
-              )}
-              {item.attributeBonuses &&
-                Object.entries(item.attributeBonuses).map(([attr, bonus]) => (
-                  <span
-                    key={attr}
-                    className="rounded bg-emerald-100 px-1.5 py-0.5 text-[10px] font-medium text-emerald-700 dark:bg-emerald-900/50 dark:text-emerald-300"
-                  >
-                    {attr.toUpperCase()}: +{bonus}
-                  </span>
-                ))}
-            </div>
-          )}
-
-          {/* Modifications */}
-          {hasCapacity && (
-            <div>
-              <div className="flex items-center justify-between mb-1.5">
-                <span className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400 dark:text-zinc-500">
-                  Mods ({cyberItem.capacityUsed || 0}/{cyberItem.capacity} cap)
-                </span>
-                {remainingCapacity > 0 && onAddEnhancement && (
-                  <button
-                    onClick={onAddEnhancement}
-                    className="flex items-center gap-0.5 text-[10px] text-amber-600 hover:text-amber-700 dark:text-amber-400"
-                  >
-                    <Plus className="h-3 w-3" />
-                    Add
-                  </button>
-                )}
-              </div>
-              {cyberItem.enhancements && cyberItem.enhancements.length > 0 ? (
-                <div className="space-y-1">
-                  {cyberItem.enhancements.map((enh, idx) => (
-                    <div
-                      key={`${enh.catalogId}-${idx}`}
-                      className="flex items-center justify-between text-xs"
-                    >
-                      <span className="text-zinc-600 dark:text-zinc-400">
-                        • {enh.name}
-                        {enh.rating && (
-                          <span className="text-zinc-400 dark:text-zinc-500"> R{enh.rating}</span>
-                        )}
-                      </span>
-                      <div className="flex items-center gap-1">
-                        <span className="text-zinc-400">¥{formatCurrency(enh.cost)}</span>
-                        {onRemoveEnhancement && (
-                          <button
-                            onClick={() => onRemoveEnhancement(idx)}
-                            className="p-0.5 text-zinc-400 hover:text-red-500"
-                          >
-                            <X className="h-3 w-3" />
-                          </button>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              ) : (
-                <p className="text-[10px] text-zinc-400 dark:text-zinc-500 italic">None</p>
-              )}
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// =============================================================================
-// CYBERLIMB AUGMENTATION ITEM COMPONENT
-// =============================================================================
-
-function CyberlimbAugmentationItem({
-  item,
-  onRemove,
-  onAddEnhancement,
-  onRemoveEnhancement,
-  onAddAccessory,
-  onRemoveAccessory,
-  onAddWeapon,
-  onRemoveWeapon,
-}: {
-  item: CyberlimbItem;
-  onRemove: () => void;
-  onAddEnhancement?: () => void;
-  onRemoveEnhancement?: (enhancementIndex: number) => void;
-  onAddAccessory?: () => void;
-  onRemoveAccessory?: (accessoryIndex: number) => void;
-  onAddWeapon?: () => void;
-  onRemoveWeapon?: (weaponIndex: number) => void;
-}) {
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  // Calculate derived values - handle missing base values gracefully
-  // Cyberlimbs have base 3 STR/AGI by SR5 rules
-  // Formula: base (3) + custom (default 0) + enhancement rating
-  const baseStr = item.baseStrength ?? 3;
-  const baseAgi = item.baseAgility ?? 3;
-  const customStr = item.customStrength ?? 0;
-  const customAgi = item.customAgility ?? 0;
-
-  // Find enhancement bonuses from the enhancements array
-  const strEnhancement = item.enhancements?.find((e) => e.enhancementType === "strength");
-  const agiEnhancement = item.enhancements?.find((e) => e.enhancementType === "agility");
-  const armorEnhancement = item.enhancements?.find((e) => e.enhancementType === "armor");
-
-  const effectiveStr = baseStr + customStr + (strEnhancement?.rating ?? 0);
-  const effectiveAgi = baseAgi + customAgi + (agiEnhancement?.rating ?? 0);
-  const armorBonus = armorEnhancement?.rating ?? 0;
-  const remainingCapacity = (item.baseCapacity || item.capacity || 0) - (item.capacityUsed || 0);
-  const totalCapacity = item.baseCapacity || item.capacity || 0;
-
-  // Count mods
-  const enhancementCount = item.enhancements?.length || 0;
-  const accessoryCount = item.accessories?.length || 0;
-  const weaponCount = item.weapons?.length || 0;
-  const totalMods = enhancementCount + accessoryCount + weaponCount;
-
-  // Format location display
-  const locationDisplay = useMemo(() => {
-    const loc = item.location.replace(/-/g, " ");
-    return loc.charAt(0).toUpperCase() + loc.slice(1);
-  }, [item.location]);
-
-  return (
-    <div>
-      {/* Compact Single-Line Header */}
-      <div className="flex items-center gap-1.5 py-2">
-        {/* Expand/Collapse Button */}
-        <button
-          onClick={() => setIsExpanded(!isExpanded)}
-          className="shrink-0 text-zinc-400 hover:text-zinc-600 dark:hover:text-zinc-300"
-        >
-          {isExpanded ? <ChevronDown className="h-4 w-4" /> : <ChevronRight className="h-4 w-4" />}
-        </button>
-
-        {/* Name */}
-        <span
-          className="text-sm font-medium text-zinc-900 dark:text-zinc-100 truncate flex-1"
-          title={item.name}
-        >
-          {item.name}
-        </span>
-
-        {/* Essence */}
-        <span className="text-sm font-medium text-cyan-600 dark:text-cyan-400 shrink-0">
-          {formatEssence(item.essenceCost)} ESS
-        </span>
-
-        {/* Cost */}
-        <span className="text-sm font-medium text-zinc-900 dark:text-zinc-100 shrink-0">
-          ¥{formatCurrency(item.cost)}
-        </span>
-
-        {/* Mod count badge (only if has mods) */}
-        {totalMods > 0 && (
-          <span className="text-[10px] text-zinc-500 dark:text-zinc-400 shrink-0">
-            [{totalMods} mod{totalMods !== 1 ? "s" : ""}]
-          </span>
-        )}
-
-        {/* Separator */}
-        <div className="mx-1 h-5 w-px bg-zinc-300 dark:bg-zinc-600" />
-
-        {/* Remove button */}
-        <button
-          onClick={onRemove}
-          className="shrink-0 rounded p-1 text-zinc-400 hover:bg-red-100 hover:text-red-600 dark:hover:bg-red-900/30 dark:hover:text-red-400"
-        >
-          <X className="h-4 w-4" />
-        </button>
-      </div>
-
-      {/* Expanded Detail View */}
-      {isExpanded && (
-        <div className="border-t border-zinc-100 dark:border-zinc-800 py-3 ml-6 space-y-3">
-          {/* Grade & Location */}
-          <div className="text-xs text-zinc-500 dark:text-zinc-400">
-            {GRADE_DISPLAY[item.grade]} • {locationDisplay} •{" "}
-            {item.appearance === "synthetic" ? "Synthetic" : "Obvious"}
-          </div>
-
-          {/* Limb Attributes - inline */}
-          <div className="flex flex-wrap gap-1.5">
-            <span className="rounded bg-emerald-100 px-1.5 py-0.5 text-[10px] font-medium text-emerald-700 dark:bg-emerald-900/50 dark:text-emerald-300">
-              STR {effectiveStr}
-            </span>
-            <span className="rounded bg-emerald-100 px-1.5 py-0.5 text-[10px] font-medium text-emerald-700 dark:bg-emerald-900/50 dark:text-emerald-300">
-              AGI {effectiveAgi}
-            </span>
-            {armorBonus > 0 && (
-              <span className="flex items-center gap-1 rounded bg-amber-100 px-1.5 py-0.5 text-[10px] font-medium text-amber-700 dark:bg-amber-900/50 dark:text-amber-300">
-                <Shield className="h-3 w-3" />+{armorBonus} Armor
-              </span>
-            )}
-          </div>
-
-          {/* Enhancements */}
-          <div>
-            <div className="flex items-center justify-between mb-1.5">
-              <span className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400 dark:text-zinc-500">
-                Enhancements ({enhancementCount})
-              </span>
-              {remainingCapacity > 0 && onAddEnhancement && (
-                <button
-                  onClick={onAddEnhancement}
-                  className="flex items-center gap-0.5 text-[10px] text-amber-600 hover:text-amber-700 dark:text-amber-400"
-                >
-                  <Plus className="h-3 w-3" />
-                  Add
-                </button>
-              )}
-            </div>
-            {item.enhancements && item.enhancements.length > 0 ? (
-              <div className="space-y-1">
-                {item.enhancements.map((enh, idx) => (
-                  <div
-                    key={`${enh.catalogId}-${idx}`}
-                    className="flex items-center justify-between text-xs"
-                  >
-                    <span className="text-zinc-600 dark:text-zinc-400">
-                      • {enh.name}
-                      <span className="text-zinc-400 dark:text-zinc-500">
-                        {" "}
-                        +{enh.rating} [{enh.capacityUsed}]
-                      </span>
-                    </span>
-                    <div className="flex items-center gap-1">
-                      <span className="text-zinc-400">¥{formatCurrency(enh.cost)}</span>
-                      {onRemoveEnhancement && (
-                        <button
-                          onClick={() => onRemoveEnhancement(idx)}
-                          className="p-0.5 text-zinc-400 hover:text-red-500"
-                        >
-                          <X className="h-3 w-3" />
-                        </button>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            ) : (
-              <p className="text-[10px] text-zinc-400 dark:text-zinc-500 italic">None</p>
-            )}
-          </div>
-
-          {/* Accessories */}
-          <div>
-            <div className="flex items-center justify-between mb-1.5">
-              <span className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400 dark:text-zinc-500">
-                Accessories ({accessoryCount})
-              </span>
-              {remainingCapacity > 0 && onAddAccessory && (
-                <button
-                  onClick={onAddAccessory}
-                  className="flex items-center gap-0.5 text-[10px] text-amber-600 hover:text-amber-700 dark:text-amber-400"
-                >
-                  <Plus className="h-3 w-3" />
-                  Add
-                </button>
-              )}
-            </div>
-            {item.accessories && item.accessories.length > 0 ? (
-              <div className="space-y-1">
-                {item.accessories.map((acc, idx) => (
-                  <div
-                    key={`${acc.catalogId}-${idx}`}
-                    className="flex items-center justify-between text-xs"
-                  >
-                    <span className="text-zinc-600 dark:text-zinc-400">
-                      • {acc.name}
-                      {acc.rating && (
-                        <span className="text-zinc-400 dark:text-zinc-500"> R{acc.rating}</span>
-                      )}
-                      <span className="text-zinc-400 dark:text-zinc-500">
-                        {" "}
-                        [{acc.capacityUsed}]
-                      </span>
-                    </span>
-                    <div className="flex items-center gap-1">
-                      <span className="text-zinc-400">¥{formatCurrency(acc.cost)}</span>
-                      {onRemoveAccessory && (
-                        <button
-                          onClick={() => onRemoveAccessory(idx)}
-                          className="p-0.5 text-zinc-400 hover:text-red-500"
-                        >
-                          <X className="h-3 w-3" />
-                        </button>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            ) : (
-              <p className="text-[10px] text-zinc-400 dark:text-zinc-500 italic">None</p>
-            )}
-          </div>
-
-          {/* Weapons */}
-          <div>
-            <div className="flex items-center justify-between mb-1.5">
-              <span className="text-[10px] font-semibold uppercase tracking-wider text-zinc-400 dark:text-zinc-500">
-                Implant Weapons ({weaponCount})
-              </span>
-              {remainingCapacity > 0 && onAddWeapon && (
-                <button
-                  onClick={onAddWeapon}
-                  className="flex items-center gap-0.5 text-[10px] text-amber-600 hover:text-amber-700 dark:text-amber-400"
-                >
-                  <Plus className="h-3 w-3" />
-                  Add
-                </button>
-              )}
-            </div>
-            {item.weapons && item.weapons.length > 0 ? (
-              <div className="space-y-1">
-                {item.weapons.map((wpn, idx) => (
-                  <div
-                    key={`${wpn.catalogId}-${idx}`}
-                    className="flex items-center justify-between text-xs"
-                  >
-                    <span className="text-zinc-600 dark:text-zinc-400">
-                      • {wpn.name}
-                      <span className="text-zinc-400 dark:text-zinc-500">
-                        {" "}
-                        ({wpn.damage}/{wpn.ap}AP) [{wpn.capacityUsed}]
-                      </span>
-                    </span>
-                    <div className="flex items-center gap-1">
-                      <span className="text-zinc-400">¥{formatCurrency(wpn.cost)}</span>
-                      {onRemoveWeapon && (
-                        <button
-                          onClick={() => onRemoveWeapon(idx)}
-                          className="p-0.5 text-zinc-400 hover:text-red-500"
-                        >
-                          <X className="h-3 w-3" />
-                        </button>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            ) : (
-              <p className="text-[10px] text-zinc-400 dark:text-zinc-500 italic">None</p>
-            )}
-          </div>
-
-          {/* Capacity summary */}
-          <div className="text-[10px] text-zinc-400 dark:text-zinc-500">
-            Capacity: {item.capacityUsed || 0}/{totalCapacity} used ({remainingCapacity} remaining)
-          </div>
-        </div>
-      )}
-    </div>
-  );
 }
 
 // =============================================================================
